addressing PR comments (includes reverting ledger.h splitting changes)
